Cheatsheet added to index and original .ipynb moved to code folder
The cheatsheet has been added to the main project directory as a .md file.

The original .ipynb file is in the code folder.

A new folder has been created to store documents, to enable students to print out more easily the non-code guides (such as the cheatsheet I have just made).