Cheatsheet added, and changes to the directories and their contents (#3)
* Updated README

* The guide is incomplete, so don't rely on it just yet.

* Added formatting to the document and expanded a little the explanation

* Double-checking the up-to-date document has been uploaded

* Reworded the motivation section

* I moved the header tags and removed the HTML at the end of the document. I also added some notes for future work.

* Jupyter Cheatsheet

A cheatsheet for Jupyter Notebook and Jupyter Lab beginners. It explains the most essential commands for quick programming.

* Removed Runtime Information from Script

* Cheatsheet added to index and original .ipynb moved to code folder

The cheatsheet has been added to the main project directory as a .md file.

The original .ipynb file is in the code folder.

A new folder has been created to store documents, to enable students to print out more easily the non-code guides (such as the cheatsheet I have just made).